# Optimize the performance of Circuit s_external mehod