This PR aims to improve the performance of s_external method. Related to https://github.com/scikit-rf/scikit-rf/issues/1011 https://github.com/scikit-rf/scikit-rf/pull/1070

This optimization accelerates the method by using `np.linalg.solve(A, B)` instead of `np.inv(A) @ B`. The benefits of this change become more pronounced as the circuit size increases. Benchmark code can be found in https://github.com/scikit-rf/scikit-rf/pull/1080, and the results are as follows:
master branch benchmark:
```bash
Total time in Easy example: 3.8792 ms, min: 3.7143 ms, std: 0.2074 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 8.1559 ms, min: 7.9134 ms, std: 0.2884 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 48.4151 ms, min: 43.9246 ms, std: 2.5155 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 48.1889 ms, min: 47.2999 ms, std: 0.8018 ms
Standard (47) and Optimized (13) results are equal: True
```
this branch benchmark:
```bash
Total time in Easy example: 3.5822 ms, min: 3.3584 ms, std: 0.1777 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 7.8416 ms, min: 7.4914 ms, std: 0.2966 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 41.4574 ms, min: 37.2748 ms, std: 3.1524 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 48.1995 ms, min: 47.0262 ms, std: 1.0177 ms
Standard (47) and Optimized (13) results are equal: True
```
If we set the `auto_reduce=False`, thus disable optimization to fully calculate the s-parameters of the entire circuit without pre-simplification, the results are as follows:
master branch benchmark:
```bash
Total time in Easy example: 4.3446 ms, min: 3.9978 ms, std: 0.4021 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 11.2693 ms, min: 10.0007 ms, std: 0.9517 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 328.7104 ms, min: 314.2743 ms, std: 5.3674 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 132.3545 ms, min: 126.2543 ms, std: 3.2932 ms
Standard (47) and Optimized (13) results are equal: True
```
this branch benchmark:
```bash
Total time in Easy example: 3.7349 ms, min: 3.4740 ms, std: 0.2870 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 10.4233 ms, min: 9.3174 ms, std: 0.5318 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 231.0316 ms, min: 224.7889 ms, std: 3.8924 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 83.1854 ms, min: 78.1748 ms, std: 2.6112 ms
Standard (47) and Optimized (13) results are equal: True
```
It can be seen that when `auto_reduce=False` is not performed, the speed of complex model and HFSS model with **more than 40 nodes** can be improved by nearly **30%**.

If necessary, I can submit a performance comparison between `np.linalg.solve(A, B)` and `np.inv(A) @ B` for different matrix sizes to clearly demonstrate the optimization effect.

Additionally, according to my current profiling, after applying this PR, the performance bottleneck will be concentrated on `np.matmul(t, x, out=t)`, that is to say `self.C @ self.X`. At present, I have no further ideas for optimizing matrix multiplication.

Your feedback and suggestions on this optimization are welcome!