@Asachoo 2024-06-21
Regarding `self.C @ self.X`, I have some preliminary ideas and would like to gather feedback from everyone.

`Network.X` is a block diagonal matrix, and needs to compute the individual intersection matrices `[X]_k` before assembling Network.X. As a result, the assembly of `Network.X` and the computation of `self.C @ self.X` have become two significant numerical bottlenecks in the `Circuit.s_external()` method.

https://github.com/scikit-rf/scikit-rf/blob/d806cee5c41b7d8e8697eb4c61527c100d40467c/skrf/circuit.py#L687-L696
https://github.com/scikit-rf/scikit-rf/blob/d806cee5c41b7d8e8697eb4c61527c100d40467c/skrf/circuit.py#L861-L863

So, why not try to directly use the individual intersection matrices [X]_k for the calculations? This approach can avoid the assembly of `Network.X` and circumvent the computation of `self.C @ self.X`, replacing it with a for loop that computes `self.C[:, :, [X]_k_indices] @ [X]_k`.

To explore this idea, I have modified a new `s_external()` method implementing this approach:
```python
    @property
    def _Xks(self) -> tuple[np.ndarray, ...]:
        """
        Return a tuple of individual intersection matrices [X]_k of the circuit.

        Each matrix [X]_k is assembled based on the connections in the circuit and
        returned as a tuple of numpy arrays.

        Returns
        -------
        Xks : tuple[np.ndarray, ...]
            A tuple containing the individual intersection matrices [X]_k for each
        connection in the circuit.

        Note
        ----
        This function avoids the numerical bottleneck associated with creating a 
        large block diagonal matrix by returning the individual matrices directly.
        These matrices can be used separately or assembled into a block diagonal 
        matrix as needed.
        """
        Xks = tuple([self._Xk(cnx) for cnx in self.connections])
        return Xks

    @property
    def s_external(self) -> np.ndarray:
        """
        Return the scattering parameters for the external ports.

        Returns
        -------
        S : :class:`numpy.ndarray`
            Scattering parameters of the circuit for the external ports.
            Shape `f x nb_ports x nb_ports`
        """
        # The external S-matrix is the submatrix corresponding to external ports:
        # port_indexes = self.port_indexes
        # a, b = np.meshgrid(port_indexes, port_indexes, indexing='ij')
        # S_ext = self.s[:, a, b]

        # Instead of calculating all S-parameters and taking a submatrix,
        # the following faster approach only calculates external the S-parameters
        # from block-matrix operations.
        # generate index lists of internal and external ports
        ext_idxs = np.array(self.port_indexes)
        in_idxs = np.setdiff1d(np.arange(self.dim), ext_idxs)

        ext_l, in_l = len(ext_idxs), len(in_idxs)

        # generate index slices for each sub-matrices
        idx_a, idx_b, idx_c, idx_d = (
            np.repeat(i, l, axis=1)
            for i, l in (
                (ext_idxs[:, None], ext_l),
                (ext_idxs[:, None], in_l),
                (in_idxs[:, None], ext_l),
                (in_idxs[:, None], in_l),
            )
        )

        # sub-matrices index, Matrix = [[A, B], [C, D]]]
        A_idx = (slice(None), idx_a, idx_a.T)
        B_idx = (slice(None), idx_b, idx_c.T)
        C_idx = (slice(None), idx_c, idx_b.T)
        D_idx = (slice(None), idx_d, idx_d.T)

        # Get the buffer of global matrix C (stored in t) sub-matrices tuple of X (Xks)
        t, Xks = self.C, self._Xks

        # Iterate calculate the self.C @ self.X as a block diagonal matrix
        cur_idx = 0
        for xk in Xks:
            l = xk.shape[-1]
            t[:, :, cur_idx : cur_idx + l] @= -xk
            cur_idx = cur_idx + l

        diag_indices = np.arange(t.shape[-1])
        t[:, diag_indices, diag_indices] += 1.0 + 0.0j

        # Get the sub-matrices of inverse of intermediate temporary matrix t
        tmp_mat = np.linalg.solve(t[D_idx], t[C_idx])

        def _find_pos(a: list[int], n: int) -> tuple[int, int, int, int]:
            """
            Find the position and value for given index in a list of lengths.

            Parameters
            ----------
            lengths : list[int]
                List of lengths.
            index : int
                Index to find the position for.

            Returns
            -------
            tuple[int, int, int, int] : The position and value for given index
            in a list of lengths. The first element is the Network index of the 
            Networks port number tuple, the second element is the port index of
            the selected Network, the third and forth elements are the start and
            end index of the selected port in the concatenated S-matrix.

            Examples
            --------
            >>> a = [2, 4, 2]
            >>> n = 5
            >>> _find_pos(a, n)
            >>> (1, 3, 2, 6)
            """
            cumulative_sum = 0
            for idx, value in enumerate(a):
                cumulative_sum += value
                if cumulative_sum > n:
                    remainder = value - (cumulative_sum - n)
                    previous_cumulative_sum = cumulative_sum - value
                    return idx, remainder, previous_cumulative_sum, cumulative_sum

            raise ValueError("n is out of range")

        # Get lengths of each sub-matrices
        Xk_lengths = [Xk.shape[-1] for Xk in Xks]

        # Initialize `xs` with zeros and fill it with appropriate values
        xs = np.zeros((len(self.frequency), ext_l, ext_l + in_l), dtype="complex")
        for i, pi in enumerate(ext_idxs):
            n, j, p, q = _find_pos(Xk_lengths, pi)
            xs[:, i, p:q] = Xks[n][:, j, :]

        # Compute the external S-parameters using block-matrix operations
        S_ext = (xs[:, :, ext_idxs] - xs[:, :, in_idxs] @ tmp_mat) @ np.linalg.inv(
            t[A_idx] - t[B_idx] @ tmp_mat
        )

        S_ext = s2s(S_ext, self.port_z0, S_DEF_DEFAULT, 'traveling')
        return S_ext  # shape (nb_frequency, nb_ports, nb_ports)
```
Benchmark code can be found in https://github.com/scikit-rf/scikit-rf/pull/1080, and the results are as follows:
master branch benchmark:
```bash
Total time in Easy example: 3.9358 ms, min: 3.7699 ms, std: 0.2697 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 8.0533 ms, min: 7.8361 ms, std: 0.2157 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 48.7290 ms, min: 44.6756 ms, std: 2.5768 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 49.3408 ms, min: 47.1125 ms, std: 1.8269 ms
Standard (47) and Optimized (13) results are equal: True
```
this modified approach:
```bash
Total time in Easy example: 3.5033 ms, min: 3.2692 ms, std: 0.1748 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 7.6679 ms, min: 7.1812 ms, std: 0.3180 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 26.7056 ms, min: 24.3053 ms, std: 2.2692 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 50.1026 ms, min: 46.3486 ms, std: 2.6062 ms
Standard (47) and Optimized (13) results are equal: True
```
If we set the auto_reduce=False, thus disable optimization to fully calculate the s-parameters of the entire circuit without pre-simplification, the results are as follows:
master branch benchmark:
```bash
Total time in Easy example: 4.0425 ms, min: 3.7982 ms, std: 0.1712 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 10.5094 ms, min: 9.7990 ms, std: 0.6346 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 323.2934 ms, min: 308.0015 ms, std: 4.6130 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 131.2161 ms, min: 120.9514 ms, std: 3.6630 ms
Standard (47) and Optimized (13) results are equal: True
```
this modified approach:
```bash
Total time in Easy example: 3.6603 ms, min: 3.3412 ms, std: 0.2342 ms
Standard (10) and Optimized (10) results are equal: True

Total time in Simple example: 7.7435 ms, min: 7.2000 ms, std: 0.3990 ms
Standard (18) and Optimized (9) results are equal: True

Total time in Complex example: 224.0683 ms, min: 219.6231 ms, std: 2.4418 ms
Standard (80) and Optimized (26) results are equal: True

Total time in HFSS example: 65.5429 ms, min: 59.9597 ms, std: 3.8912 ms
Standard (47) and Optimized (13) results are equal: True
```