@Asachoo 2024-06-21
Unfortunately, the modified method is more complex and harder to understand, but perhaps that's the cost of optimization :)

While it has not shown significant speed improvements in some cases, it has achieved a **20%** to **30%** performance boost in others [Complex example with `auto_reduce=True`: 37.2748 ms -> 24.3053 ms, and HFSS example with `auto_reduce=False` 78.1748 ms -> 59.9597 ms]  (comparing my local branch with this new method to the current PR).

I suspect that the performance variability stems from the trade-off between the decreased efficiency of the for loop and the increased efficiency from reducing matrix sizes. Additionally, I believe using technologies like numba@jit or Cython could mitigate the efficiency issues of the for loop.

Since this method does not always show a significant performance improvement, I would like to hear your suggestions on whether to push this change to the PR immediately or to make further modifications.